requirements.txt
torch
torchvision
Pillow
exif
streamlit
pandas